+++ test/runtests.jl
@@ -16,7 +16,7 @@ PowerModels.silence()
 pms_path = joinpath(dirname(pathof(PowerModels)), "..")
 
 # default setup for solvers
-ipopt_solver = JuMP.with_optimizer(Ipopt.Optimizer, tol=1e-6, print_level=0)
+ipopt_solver = JuMP.optimizer_with_attributes(Ipopt.Optimizer, "tol"=>1e-6, "print_level"=>0)
 
 
 # this will work because PowerModels is a dependency